% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/backbones_evaluate.R
\name{backbones_evaluate}
\alias{backbones_evaluate}
\title{Evaluate one or multiple backbones according to various metrics.}
\usage{
backbones_evaluate(BL, G = NULL)
}
\arguments{
\item{BL}{A backbone or list of backbone objects.}

\item{G}{(Optional) The original graph G, assumed to be equal for all backbones.
I
Standard is NULL.}
}
\value{
A data frame with one row per backbone, evaluating it according to the following metrics:
               - V = fraction of included nodes
               - E = fraction of included edges
               - R = goodness of fit
               - sigma = smoothness
               - cor_act = average commute time preservation
               - leaves = number of leaves in the backbone
}
\description{
Evaluate one or multiple backbones according to various metrics.
}
