% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/boundary_coefficient.R
\name{boundary_coefficient}
\alias{boundary_coefficient}
\title{Compute the boundary coefficients of a weighted graph.}
\usage{
boundary_coefficient(G, constrain_mem = FALSE)
}
\arguments{
\item{G}{The graph of which boundary coefficients are to be computed.}

\item{constrain_mem}{(Optional) If TRUE and pairD is missing, the entire distance matrix on G as is computed first, from which the hop-k-approximation is extracted afterwards.
If FALSE, entries that are not needed in pairD are never computed, requiring less storage.
Standard is FALSE, which leads to faster results with our current implementation.}
}
\value{
A vector of |V(G)| boundary coefficients.
}
\description{
Compute the boundary coefficients of a weighted graph.
}
