% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_for_cycles.R
\name{check_for_cycles}
\alias{check_for_cycles}
\title{Check for any gaps in the obtained backbone by means of persistent homology.
As R is currently not ideal for persistent homology computation, this method can quikly lead to memory issues.}
\usage{
check_for_cycles(backbone, G = NULL)
}
\arguments{
\item{backbone}{A backbone object for which gaps need to be identified.}

\item{G}{(Optional) the original graph on top of the backbone.
Must be included if not contained in the backbone object.
Standard is NULL.}
}
\value{
The backbone object (list) with the following (new) entries:
             - diagram: the 0-dimensional and 1-dimensional persistence diagrams in a matrix.
             - repCycles: list of edges that represent the 1-dimensional holes in diagram.
}
\description{
Check for any gaps in the obtained backbone by means of persistent homology.
As R is currently not ideal for persistent homology computation, this method can quikly lead to memory issues.
}
