% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/decompose.R
\name{decompose}
\alias{decompose}
\title{Decompose a bacbone according to branches, i.e., maximal paths with no multifurcations.}
\usage{
decompose(backbone, G = NULL, prob = TRUE)
}
\arguments{
\item{backbone}{A backbone object.}

\item{G}{(Optional) the original graph on top of the backbone.
Must be included if not contained in the backbone object.
Standard is NULL.}

\item{prob}{(Optional & if assign) whether to assign nodes in G to paths according to a probability measure
if FALSE, then nodes are assigned to exactly one path through a one-hot-encoding.
Standard is TRUE.}
}
\value{
The backbone object (list) with the following (new) entries:
               - branch: a vector of integers clustering the edges of backbone into paths.
               - prob: a sparse n x k matrix, where n = |V(G)| and k is the number of branches in the backbone, clustering points of G according to branches of the backbone.
               - Palette: a vector of hex code colors, one for each branch.
               - Col: a vector of |V(G)| hex code colors, interpolating between the palette colors according to prob.
}
\description{
Decompose a bacbone according to branches, i.e., maximal paths with no multifurcations.
}
