% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_elbow.R
\name{get_elbow}
\alias{get_elbow}
\title{Compute the elbows for a set of CLOF cost functions (one per component) by minimizing the second-order finite differentiates.}
\usage{
get_elbow(C)
}
\arguments{
\item{C}{CLOF cost function, as stored in a backbone object.}
}
\value{
data.frame where each row contains the elbow (leaves + cost) for the corresponding component.
}
\description{
Compute the elbows for a set of CLOF cost functions (one per component) by minimizing the second-order finite differentiates.
}
