% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_leave_combs.R
\name{get_leave_combs}
\alias{get_leave_combs}
\title{For a given number of leaves, compute all possible leave configurations over a number of components.}
\usage{
get_leave_combs(c, l)
}
\arguments{
\item{c}{Number of components.}

\item{l}{Number which the sum of leaves over all components must equal.}
}
\value{
A matrix with c columns, where each row is a possible combination of leaves over the c components.
}
\description{
For a given number of leaves, compute all possible leave configurations over a number of components.
}
