% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_new_leaves.R
\name{get_new_leaves}
\alias{get_new_leaves}
\title{Get a new backbone with a different amount of leaves}
\usage{
get_new_leaves(
  backbone,
  leaves,
  G = NULL,
  assign = FALSE,
  prob = TRUE,
  stdize = FALSE
)
}
\arguments{
\item{backbone}{A backbone object to be converted to a new backbone object with a different amount of leaves}

\item{leaves}{Vector of integers, denoting the number of leaves to be included for each component in the pine.
Must have length 1 or length equal to the number of connected components included in backbone.
If length 1, we optimize for the total number of leaves equal to the number specified.
If length equal to the number of connected components, we extract the specified number of leaves for each component.}

\item{G}{(Optional) the original graph on top of the backbone.
Must be included if not contained in the backbone object, and either assign is TRUE or the backbone already had a branch assignment.
Standard is NULL.}

\item{assign}{(Optional) Whether to produce a branch assignment of the resulting new backbone.
Will always be treated as TRUE if the backbone already had a branch assignment.
Standard is FALSE.}

\item{prob}{(Optional) Whether to assign nodes in G to paths according to a 'probability'.
If FALSE, then nodes are assigned to exactly one path through a one-hot-encoding.
Standard is TRUE.}

\item{stdize}{(Optional) Whether to standardize the cost of each component according to its full cost.
Standard is FALSE, which may add bias towards representing larger components in the backbone.}
}
\value{
a backbone object with the new number of leaves
}
\description{
Get a new backbone with a different amount of leaves
}
