% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graph_complex.R
\name{graph_complex}
\alias{graph_complex}
\title{Convert an igraph object to a simplicial complex.}
\usage{
graph_complex(G)
}
\arguments{
\item{G}{The input graph.}
}
\value{
A simplicial complex S, compatible with the TDA library in R.
}
\description{
Convert an igraph object to a simplicial complex.
}
