% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graph_filtration.R
\name{graph_filtration}
\alias{graph_filtration}
\title{Compute a filtration from a vertex-valued defined over a graph G.}
\usage{
graph_filtration(S, f, sublevel = TRUE)
}
\arguments{
\item{S}{A graph represented by either an igraph or cmplx object.}

\item{f}{A vertex-valued function defined over the nodes of G.}

\item{sublevel}{(Optional) If FALSE, superlevel filtration is computed instead.
Standard is TRUE.}
}
\value{
A filtration defined by f on G, compatible with the TDA library in R.
}
\description{
Compute a filtration from a vertex-valued defined over a graph G.
}
