% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graph_persistence.R
\name{graph_persistence}
\alias{graph_persistence}
\title{Compute zero-dimensional persistence from a vertex-valued defined over a graph G.}
\usage{
graph_persistence(FltFun = NULL, S = NULL, f = NULL, sublevel = TRUE)
}
\arguments{
\item{FltFun}{(Optional) Filtration of which to compute persistence.
If missing, both the graph and function values must be provided.}

\item{S}{(Optional) A graph represented by either an igraph or cmplx object.}

\item{f}{(Optional) A vertex-valued function defined over the nodes of G.}

\item{sublevel}{(Optional) Only used if FltFun is missing.
If FALSE, superlevel filtration is computed instead.
Standard is TRUE.}
}
\value{
The result of the persistent homology computation for the specified filtration through the TDA library in R.
}
\description{
Compute zero-dimensional persistence from a vertex-valued defined over a graph G.
}
