% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kNNgraph.R
\name{kNNgraph}
\alias{kNNgraph}
\title{Construct an undirected kNN graph by connecting each point to its k closest neighbors.}
\usage{
kNNgraph(X, k = 10)
}
\arguments{
\item{X}{A numeric data.frame or matrix containing observations, or a distance object containing distances between pairs of observations.}

\item{k}{Number of neighbors in the kNN graph.
Standard is 10.}
}
\value{
Undirected k-nearest neighbor graph G (igraph object).
         Note that the order of nodes in X and G may differ.
         Rownames in X correspond to vertex names in G.
}
\description{
Construct an undirected kNN graph by connecting each point to its k closest neighbors.
}
