% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/normalized_centrality.R
\name{normalized_centrality}
\alias{normalized_centrality}
\title{Compute the normalized centrality for each node in a given graph G.}
\usage{
normalized_centrality(G)
}
\arguments{
\item{G}{The input graph.}
}
\value{
A numeric vector of length(V(G)) containing the degrees of centrality, computed per component.
}
\description{
Compute the normalized centrality for each node in a given graph G.
}
