% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/project_on_backbone.R
\name{project_on_backbone}
\alias{project_on_backbone}
\title{Project a given graph G on a backbone subgraph G.}
\usage{
project_on_backbone(G, B, dG = NULL)
}
\arguments{
\item{G}{igraph object.}

\item{B}{Backbone subgraph of G.}

\item{dG}{(Optional) pairwise distance matrix of G.
Computed if not provided.}
}
\value{
The projection of G on B (list/igraph object), with two added entries:
              - addedV: the set difference between the nodes of G and B.
              - connectTo: for each node in addedV, the point in B on which it is projected.
}
\description{
Project a given graph G on a backbone subgraph G.
}
