% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/VRgraph.R
\name{VRgraph}
\alias{VRgraph}
\title{Construct a Vietoris-Rips graph from by connecting each point to all points within a certain distance eps.}
\usage{
VRgraph(X, eps)
}
\arguments{
\item{X}{A numeric data.frame or matrix containing observations, or a distance object containing distances between pairs of observations.}

\item{eps}{Upper bound on edge weights/lengths.}
}
\value{
Undirected Vietoris-Rips graph G (igraph object).
        Note that the order of nodes in X and G may differ.
        Rownames in X correspond to vertex names in G.
}
\description{
Construct a Vietoris-Rips graph from by connecting each point to all points within a certain distance eps.
}
